Cleaning up
Removed a whole bunch of test and temporary files.  What remains is the
final Flownet project, as passed in as final work.

+++ flowconf.sh
@@ -15,20 +15,26 @@ then
     exit 1
 fi
 
+clear
+
 while :
 do
-
-echo -e "What would you like to do?\
+echo
+echo "Welcome to the Flownet installation/configuration wizard!"
+echo -e "What would you like to do?\n\
+\n\
 Input \"install\" for a fresh installation of Flownet\n\
 Input \"add\" to add a server to your Flownetwork (new server must have Flownet configured already)\n\
 Input \"log\" to view the current Flowlog\n\
 Input \"clear\" to clear the current Flowlog\n\
-Input \"edit\" to view and edit the local flownet.conf\n
+Input \"edit\" to view and edit the local flownet.conf (uses VIM)\n\
 Input \"quit\" to banish this wizard"
 
 if [ ! -d /home/flownet/ ]
 then
     echo "Flownet is NOT installed!!! (add, log, clear, and edit will FAIL)"
+else
+    echo "Note: A Flownet installation already exists on this system!"
 fi
 
 read INPUT
@@ -40,8 +46,8 @@ case $INPUT in
 	case $SURE in
 	    yes)
 		# Empty /home/flownet directory
-		echo "Removing the /home/flownet/ directory..."
-		rm -rf /home/flownet/ 
+		echo "Emptying the /home/flownet/ directory..."
+		rm -rf /home/flownet/* 
 		echo "...Done!"
 		echo
 
@@ -75,7 +81,8 @@ case $INPUT in
 		# Create flowlog.txt
 		echo "Creating the Flownet log file..."
 		rm -f /home/flownet/flowlog.txt >/dev/null
-		echo "FLOWNET LOG" >>flowlog.txt
+		echo "FLOWNET LOG" >>/home/flownet/flowlog.txt
+		chown flownet:flownet /home/flownet/flowlog.txt
 		echo "...Done!"
 
 		# Create flownet.sh
@@ -93,12 +100,12 @@ case $INPUT in
 
 		# Generate RSA key for local Flownet user
 		echo "Generating local RSA keys..."
-		echo "When prompted, press enter for no password:"
+		echo "When prompted, press enter for default location and no password:"
 		su flownet -c "ssh-keygen -t rsa"
 		echo "...Done!"
-		
-		echo "Flownet has now sucessfully been installed!  You should now use the edit command to configure flownet, the add command to add servers to your Flownetwork, and restart once you're ready."
 		echo
+
+		echo "Flownet was successfully installed!  You should now used the edit command to configure Flownet, and then use the add command to add servers to your network."
 		;;
 
 	    no)
@@ -129,25 +136,39 @@ case $INPUT in
 	su flownet -c "ssh flownet@${NEWSERVER} chmod 700 /home/flownet/.ssh/authorized_keys"
 	echo "...Done!"
 	echo
-	
 	;;
 
     log)
 	# View the Flowlog
-	more /home/flownet/flowlog.txt
+	if [ -f /home/flownet/flowlog.txt ]
+	then
+	    more /home/flownet/flowlog.txt
+	else
+	    echo "Flowlog does not yet exist!"
+	fi
 	;;
 
     clear)
 	# Clear the Flowlog
-	echo "Clearing the Flownet log file..."
-	rm -f /home/flownet/flowlog.txt
-	echo "FLOWNET LOG" >>flowlog.txt
-	echo "...Done!"
+	if [ -f /home/flownet/flowlog.txt ]
+	then
+	    echo "Clearing the Flownet log file..."
+	    rm -f /home/flownet/flowlog.txt
+	    echo "FLOWNET LOG" >>/home/flownet/flowlog.txt
+	    echo "...Done!"
+	else
+	    echo "Flowlog does not yet exist!"
+	fi
 	;;
 
     edit)
 	# Edit flownet.conf
-	vim /home/flownet/flownet.conf
+	if [ -f /home/flownet/flownet.conf ]
+	then
+	    vim /home/flownet/flownet.conf
+	else
+	    echo "flownet.conf does not exist yet!"
+	fi
 	;;
 
     quit)
